[test-early-flake-detection] Moved test
