+++ components/datadog/apps/logger/images/logger/main_test.go
@@ -2,9 +2,7 @@ package main
 
 import (
 	"bytes"
-	"fmt"
 	"io"
-	"math/rand/v2"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -14,20 +12,6 @@ import (
 	"testing"
 	"time"
 )
-
-func TestImFlaky(t *testing.T) {
-	fmt.Println("I'm flaky!")
-	if rand.Int32()%2 == 0 {
-		fmt.Println("I fail")
-		t.Fail()
-	} else {
-		fmt.Println("I dont fail")
-	}
-}
-
-func TestImOk(t *testing.T) {
-	fmt.Println("OK")
-}
 
 func TestPostHandler(t *testing.T) {
