+++ integration-tests/invoke_test.go
@@ -5,6 +5,7 @@ import (
 	_ "embed"
 	"flag"
 	"fmt"
+	"math/rand/v2"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -18,6 +19,20 @@ import (
 )
 
 var workingDir = flag.String("workingDir", "", "run tests in the specified directory, relative to the root of the repository. Defaults to the root of the repository.")
+
+func TestImFlaky(t *testing.T) {
+	fmt.Println("I'm flaky!")
+	if rand.Int32()%2 == 0 {
+		fmt.Println("I fail")
+		t.Fail()
+	} else {
+		fmt.Println("I dont fail")
+	}
+}
+
+func TestImOk(t *testing.T) {
+	fmt.Println("OK")
+}
 
 func TestInvokes(t *testing.T) {
 	// Set working directory
